Development

+++ server/api/formulas/[id].delete.ts
@@ -1,4 +1,7 @@
+import { requireAdminLikeUser } from '../../utils/rbac';
+
 export default defineEventHandler(async (event) => {
+  requireAdminLikeUser(event)
 
   //delete formula
   await prisma.formulas.delete({
